requirements.txt
fastapi
uvicorn
python-multipart
openai-whisper
transformers
torch
pillow
pyttsx3
python-jose
pydantic
numpy
python-dotenv